Rating button and font cleanup
I added the rating button, which ofc has no real functionality here, but it looks mostly like it does on the main site.

I also adjusted the font family for the logo so it has a proper thematic backup. I can't host the page with Bauhaus LT but the font is a free download. I included it the files so others can use it locally.